pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
]

[project]
name = "aa-permission-management"
description = "Simple permission managewment plugin for Alliance Auth"
readme = "README.md"
keywords = [
    "allianceauth",
    "management",
    "permission",
]
license = { file = "LICENSE" }
authors = [
    { name = "Peter Pfeufer", email = "develop@ppfeufer.de" },
]
requires-python = ">=3.10,<3.14"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dynamic = [
    "version",
]
dependencies = [
    "allianceauth>=4.12,<5",
]
optional-dependencies.tests-allianceauth-latest = [
    "coverage",
    "django-webtest",
    "requests-mock",
]
urls.Changelog = "https://github.com/ppfeufer/aa-permission-management/blob/master/CHANGELOG.md"
urls.Codecov = "https://codecov.io/gh/ppfeufer/aa-permission-management"
urls.Discord = "https://discord.gg/fjnHAmk"
urls.Documentation = "https://github.com/ppfeufer/aa-permission-management/blob/master/README.md"
urls.Donation = "https://ko-fi.com/ppfeufer"
urls.GitHub = "https://github.com/ppfeufer/aa-permission-management"
urls.Tracker = "https://github.com/ppfeufer/aa-permission-management/issues"
urls.Translation = "https://weblate.ppfeufer.de/projects/alliance-auth-apps/aa-permission-management/"

[tool.hatch.build]
include = [
    "/aa_permission_management",
]

[tool.hatch.version]
path = "aa_permission_management/__init__.py"
